[refactor]: application.properties에 파일 경로 세팅
