+++ src/main/java/app/library/management/config/Configuration.java
@@ -1,6 +1,6 @@
 package app.library.management.config;
 
-import app.library.management.config.util.BookRepositoryFactory;
+import app.library.management.config.factory.BookRepositoryFactory;
 import app.library.management.core.domain.util.BookStatusManager;
 import app.library.management.core.domain.util.SchedulerManager;
 import app.library.management.infra.port.Request;
