+++ src/main/resources/application.properties
@@ -1,0 +1,1 @@
+book.repository.path=src/main/java/app/library/management/core/repository/file/Book.json
